+++ packages/ui/components/VisualEditor/SchemaProperty.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import _ from 'lodash';
 import SchemaObject from './SchemaObject';
 import PropertyControls from './PropertyControls';
-import { RequiredIcon, NotRequiredIcon } from './icons';
-import { TrashIcon } from '../components/icons'
+import { RequiredIcon, NotRequiredIcon } from '../icons';
+import { TrashIcon } from '../icons'
 
 interface SchemaPropertyProps {
     name: string;
