+++ packages/ui/components/index.tsx
@@ -13,7 +13,7 @@ export * from './Sidebar'
 export * from './SlideOver'
 export * from './Toolbar'
 export * from './AppCard'
-export * from './Example'
-export * from './CodeEditor'
+export * from './VisualEditor/Example'
+export * from './VisualEditor/CodeEditor'
 export * from './VisualEditor'
 export { default as Tooltip } from './Tooltip'
